+++ chibi/services/bot.py
@@ -11,7 +11,8 @@ from telegram import (
 from telegram.ext import ContextTypes
 
 from chibi.config import application_settings, gpt_settings
-from chibi.services.gpt import api_key_is_valid
+from chibi.constants import SupportedProviders
+from chibi.schemas.app import ChatResponseSchema
 from chibi.services.user import (
     check_history_and_summarize,
     generate_image,
@@ -26,17 +27,18 @@ from chibi.utils import (
     api_key_is_plausible,
     chat_data,
     download_image,
+    get_provider_class_by_name,
     get_telegram_chat,
     get_telegram_message,
     get_telegram_user,
     handle_gpt_exceptions,
     send_gpt_answer_message,
     send_message,
-    user_can_use_gpt4,
     user_data,
 )
 
 
+@handle_gpt_exceptions
 async def handle_model_selection(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
     query = update.callback_query
     if not query:
@@ -63,6 +65,7 @@ async def handle_prompt(update: Update, context: ContextTypes.DEFAULT_TYPE) -> N
         prompt = prompt.replace("/ask", "", 1).strip()
 
     prompt_to_log = prompt.replace("\r", " ").replace("\n", " ")
+
     logger.info(
         f"{user_data(update)} sent a new message in the {chat_data(update)}"
         f"{': ' + prompt_to_log if application_settings.log_prompt_data else ''}"
@@ -74,17 +77,20 @@ async def handle_prompt(update: Update, context: ContextTypes.DEFAULT_TYPE) -> N
         await context.bot.send_chat_action(chat_id=telegram_chat.id, action=constants.ChatAction.TYPING)
         await asyncio.sleep(2.5)
 
-    gpt_answer, usage = await get_gtp_chat_answer_task
+    chat_response: ChatResponseSchema = await get_gtp_chat_answer_task
     # usage_message = (
     #     f"Tokens used: {str(usage.get('total_tokens', 'n/a'))} "
     #     f"({str(usage.get('prompt_tokens', 'n/a'))} prompt, "
     #     f"{str(usage.get('completion_tokens', 'n/a'))} completion)"
     # )
-    answer_to_log = gpt_answer.replace("\r", " ").replace("\n", " ")
+    answer_to_log = chat_response.answer.replace("\r", " ").replace("\n", " ")
     logged_answer = f"Answer: {answer_to_log}" if application_settings.log_prompt_data else ""
 
-    logger.info(f"{user_data(update)} got GPT answer in the {chat_data(update)}. {logged_answer}")
-    await send_gpt_answer_message(gpt_answer=gpt_answer, update=update, context=context)
+    logger.info(
+        f"{user_data(update)} got {chat_response.provider} ({chat_response.model}) answer in "
+        f"the {chat_data(update)}. {logged_answer}"
+    )
+    await send_gpt_answer_message(gpt_answer=chat_response.answer, update=update, context=context)
     history_is_summarized = await check_history_and_summarize(user_id=telegram_user.id)
     if history_is_summarized:
         logger.info(f"{user_data(update)}: history successfully summarized.")
@@ -100,7 +106,7 @@ async def handle_reset(update: Update, context: ContextTypes.DEFAULT_TYPE) -> No
 
 
 @handle_gpt_exceptions
-async def handle_image_generation(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+async def handle_image_generation(update: Update, context: ContextTypes.DEFAULT_TYPE, prompt: str) -> None:
     telegram_user = get_telegram_user(update=update)
     telegram_chat = get_telegram_chat(update=update)
     telegram_message = get_telegram_message(update=update)
@@ -115,15 +121,6 @@ async def handle_image_generation(update: Update, context: ContextTypes.DEFAULT_
                 f"Sorry, you have reached your monthly images generation limit "
                 f"({gpt_settings.image_generations_monthly_limit}). Please, try again later."
             ),
-        )
-        return None
-
-    prompt = telegram_message.text.replace("/imagine", "", 1).strip()
-    if not prompt:
-        await context.bot.send_message(
-            chat_id=telegram_chat.id,
-            reply_to_message_id=telegram_message.message_id,
-            text="Please provide some description (e.g. /imagine cat)",
         )
         return None
 
@@ -154,45 +151,52 @@ async def handle_image_generation(update: Update, context: ContextTypes.DEFAULT_
             f"due to exception: {e}. Trying to send if via text message..."
         )
         await send_message(update=update, context=context, text="\n".join(image_urls), disable_web_page_preview=False)
+    if application_settings.log_prompt_data:
+        logger.info(f"{user_data(update)} got a successfully generated image(s): {image_urls}")
 
 
-async def handle_openai_key_set(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+async def handle_api_key_set(update: Update, context: ContextTypes.DEFAULT_TYPE, provider: SupportedProviders) -> None:
     telegram_user = get_telegram_user(update=update)
     telegram_chat = get_telegram_chat(update=update)
     telegram_message = get_telegram_message(update=update)
-    logger.info(f"{telegram_user.name} provides ones API Key.")
+    logger.info(f"{telegram_user.name} provides {provider.value} API Key.")
 
-    if not telegram_message.text:
+    prompt = telegram_message.text
+    if not prompt:
         return None
 
-    api_key = telegram_message.text.replace("/set_openai_key", "", 1).strip()
+    api_key = prompt.strip()
     error_msg = (
-        "Sorry, but incorrect API key provided. You can find your API key at "
-        "https://platform.openai.com/account/api-keys"
+        "Sorry, but API key you have provided does not seem correct. You can find your API key at:\n "
+        "https://platform.openai.com/account/api-keys\n"
+        "https://console.anthropic.com/settings/keys\n"
+        "https://console.mistral.ai/api-keys"
     )
     if not api_key_is_plausible(api_key=api_key):
         await send_message(update=update, context=context, text=error_msg)
         logger.warning(f"{user_data(update)} provided improbable key.")
-        return
+        return None
 
-    if not await api_key_is_valid(api_key=api_key):
+    provider_class = get_provider_class_by_name(provider_name=provider)
+
+    if not await provider_class(token=api_key).api_key_is_valid():
         await send_message(update=update, context=context, text=error_msg)
-        logger.warning(f"{user_data(update)} provided incorrect API key.")
-        return
+        logger.warning(f"{user_data(update)} provided invalid key.")
+        return None
 
-    await set_api_key(user_id=telegram_user.id, api_key=api_key)
-    msg = (
-        "Your OpenAI API Key successfully set, my functionality unlocked! 🦾\n\n"
-        "Now you also may check available models in /menu."
-    )
+    await set_api_key(user_id=telegram_user.id, api_key=api_key, provider=provider)
+    msg = "Your API Key successfully set! 🦾\n\n" "Now you may check available models in /models."
     await send_message(update=update, context=context, reply=False, text=msg)
-    await context.bot.delete_message(chat_id=telegram_chat.id, message_id=telegram_message.message_id)
-    logger.info(f"{user_data(update)} successfully set up OpenAPI Key.")
+    try:
+        await context.bot.delete_message(chat_id=telegram_chat.id, message_id=telegram_message.message_id)
+    except Exception:
+        pass
+    logger.info(f"{user_data(update)} successfully set up {provider.value} Key.")
 
 
+@handle_gpt_exceptions
 async def handle_available_model_options(update: Update, context: ContextTypes.DEFAULT_TYPE) -> InlineKeyboardMarkup:
     telegram_user = get_telegram_user(update=update)
-    can_use_gpt4 = user_can_use_gpt4(tg_user=telegram_user)
-    models_available = await get_models_available(user_id=telegram_user.id, include_gpt4=can_use_gpt4)
+    models_available = await get_models_available(user_id=telegram_user.id)
     keyboard = [[InlineKeyboardButton(model.upper(), callback_data=model)] for model in models_available]
     return InlineKeyboardMarkup(keyboard)