+++ chibi/utils.py
@@ -1,11 +1,10 @@
 import io
 from functools import wraps
-from typing import Any, Callable
+from typing import Any, Callable, Type
 from urllib.parse import parse_qs, urlparse
 
 import httpx
 from loguru import logger
-from openai import APIConnectionError, APIStatusError, RateLimitError
 from telegram import Chat as TelegramChat
 from telegram import Message as TelegramMessage
 from telegram import Update
@@ -15,10 +14,17 @@ from telegram.error import BadRequest
 from telegram.ext import ContextTypes
 
 from chibi.config import application_settings, gpt_settings, telegram_settings
-from chibi.services.user import get_api_key
-
-GROUP_CHAT_TYPES = [constants.ChatType.GROUP, constants.ChatType.SUPERGROUP]
-PERSONAL_CHAT_TYPES = [constants.ChatType.SENDER, constants.ChatType.PRIVATE]
+from chibi.constants import GROUP_CHAT_TYPES, PERSONAL_CHAT_TYPES, SupportedProviders
+from chibi.exceptions import (
+    NoApiKeyProvidedError,
+    NotAuthorizedError,
+    ServiceRateLimitError,
+    ServiceResponseError,
+)
+from chibi.services.providers.anthropic import Anthropic
+from chibi.services.providers.mistralai import MistralAI
+from chibi.services.providers.openai import OpenAI
+from chibi.services.providers.provider import Provider
 
 
 def get_telegram_user(update: Update) -> TelegramUser:
@@ -69,7 +75,7 @@ async def send_long_message(
     parse_mode: str | None = None,
 ) -> None:
     chunk_size = constants.MessageLimit.MAX_TEXT_LENGTH
-    text_chunks = [message[i: i + chunk_size] for i in range(0, len(message), chunk_size)]
+    text_chunks = [message[i : i + chunk_size] for i in range(0, len(message), chunk_size)]
     for chunk in text_chunks:
         await send_message(update=update, context=context, text=chunk, parse_mode=parse_mode)
 
@@ -120,14 +126,6 @@ def group_is_allowed(tg_chat: TelegramChat) -> bool:
     if not telegram_settings.groups_whitelist:
         return True
     return tg_chat.id in telegram_settings.groups_whitelist
-
-
-def user_can_use_gpt4(tg_user: TelegramUser) -> bool:
-    if gpt_settings.gpt4_enabled:
-        return True
-    if gpt_settings.gpt4_whitelist:
-        return any(identifier in gpt_settings.gpt4_whitelist for identifier in (str(tg_user.id), tg_user.username))
-    return False
 
 
 def user_interacts_with_bot(update: Update, context: ContextTypes.DEFAULT_TYPE) -> bool:
@@ -211,34 +209,52 @@ def handle_gpt_exceptions(func: Callable[..., Any]) -> Callable[..., Any]:
     async def wrapper(*args: Any, **kwargs: Any) -> Any:
         update: Update = kwargs.get("update") or args[1]
         context: ContextTypes.DEFAULT_TYPE = kwargs.get("context") or args[2]
+        error_msg_prefix = f"{user_data(update)} didn't get a GPT answer in the {chat_data(update)}"
         try:
             return await func(*args, **kwargs)
-        except APIConnectionError as e:
-            logger.error(
-                f"{user_data(update)} didn't get a GPT answer in the {chat_data(update)} because "
-                f"the server could not be reached: {e.__cause__}"
-            )
-            await send_message(
-                update=update, context=context, text="🥴Service not available at the moment. Please, try again later."
-            )
-            return None
-
-        except RateLimitError:
-            logger.error(f"{user_data(update)} reached a Rate Limit in the {chat_data(update)}.")
-            await send_message(update=update, context=context, text="🤐Rate Limit exceeded. We should back off a bit.")
-            return None
-
-        except APIStatusError as e:
-            logger.error(f"{user_data(update)} got a status code {e.status_code} response: {e.response.text}")
+        except NoApiKeyProvidedError as e:
+            logger.error(f"{error_msg_prefix}: {e}")
             await send_message(
                 update=update,
                 context=context,
-                text="😲Lol... we got an unexpected response from the OpenAI service! Please, try again a bit later.",
+                text="Oops! It looks like you didn't set the API key for this provider.",
+            )
+            return None
+
+        except NotAuthorizedError as e:
+            logger.error(f"{error_msg_prefix}: {e}")
+            await send_message(
+                update=update,
+                context=context,
+                text=(
+                    "We encountered an authorization problem when interacting with a remote service.\n"
+                    f"Please check your {e.provider} API key."
+                ),
+            )
+            return None
+
+        except ServiceResponseError as e:
+            logger.error(f"{error_msg_prefix}: {e}")
+
+            await send_message(
+                update=update,
+                context=context,
+                text=(
+                    f"😲Lol... we got an unexpected response from the {e.provider} service! \n"
+                    f"Please, try again a bit later."
+                ),
+            )
+            return None
+
+        except ServiceRateLimitError as e:
+            logger.error(f"{error_msg_prefix}: {e}")
+            await send_message(
+                update=update, context=context, text=f"🤐Rate Limit exceeded for {e.provider}. We should back off a bit."
             )
             return None
 
         except Exception as e:
-            logger.exception(f"{user_data(update)} got an error in the {chat_data(update)}: {e}")
+            logger.error(f"{error_msg_prefix}: {e!r}")
             msg = (
                 "I'm sorry, but there seems to be a little hiccup with your request at the moment 😥 Would you mind "
                 "trying again later? Don't worry, I'll be here to assist you whenever you're ready! 😼"
@@ -260,43 +276,16 @@ def api_key_is_plausible(api_key: str) -> bool:
     Returns:
         True if token provided looks like a token :) False otherwise.
     """
-
-    if len(api_key) > 55 or len(api_key) < 51:
+    if not api_key:
         return False
     if " " in api_key:
         return False
+    if len(api_key) < 30:
+        return False
+    if len(api_key) > 150:
+        return False
+
     return True
-
-
-def check_openai_api_key(func: Callable[..., Any]) -> Callable[..., Any]:
-    """Decorator checking if the specific user has an OpenAI API Key.
-
-    Args:
-        func: async function that does any OpenAI API service interaction.
-
-    Returns:
-        Wrapper function object.
-    """
-
-    async def wrapper(*args: Any, **kwargs: Any) -> Any:
-        update: Update = kwargs.get("update") or args[1]
-        context: ContextTypes.DEFAULT_TYPE = kwargs.get("context") or args[2]
-        telegram_user = get_telegram_user(update=update)
-
-        if not await get_api_key(user_id=telegram_user.id, raise_on_absence=False):
-            logger.info(f"{user_data(update)} has no OpenAI API Key. Suggesting to apply one...")
-            text = (
-                "Oops! It looks like you didn't set your OpenAI API Key yet. "
-                "Please, use /set_openai_key command, i.e.\n\n /set_openai_key sk-XXXXXXXXXXXXXXXXXXXXX\n\n"
-                "You may find your key in your user settings after creating an OpenAI account "
-                "(https://platform.openai.com/account/api-keys)."
-            )
-            await send_message(update=update, context=context, text=text, reply=False)
-            return None
-
-        return await func(*args, **kwargs)
-
-    return wrapper
 
 
 async def download_image(image_url: str) -> bytes:
@@ -308,29 +297,50 @@ async def download_image(image_url: str) -> bytes:
     return response.content
 
 
+def get_provider_class_by_name(provider_name: SupportedProviders) -> Type[Provider]:
+    match provider_name:
+        case SupportedProviders.ANTHROPIC:
+            return Anthropic
+        case SupportedProviders.MISTRALAI:
+            return MistralAI
+        case SupportedProviders.OPENAI:
+            return OpenAI
+    raise ValueError("Wrong provider name provided.")
+
+
 def log_application_settings() -> None:
-    mode = (
-        "<blue>PRIVATE (master OpenAI API Key is provided)</blue>"
-        if gpt_settings.api_key
-        else "<yellow>PUBLIC (no master OpenAI API Key is provided)</yellow>"
+    mode = "<yellow>PUBLIC</yellow>" if gpt_settings.public_mode else "<blue>PRIVATE</blue>"
+    is_set = "<green>SET</green>"
+    unset = "<red>UNSET</red>"
+    storage = "<red>REDIS</red>" if application_settings.redis else "<yellow>LOCAL</yellow>"
+    proxy = f"<blue>{telegram_settings.proxy}</blue>" if telegram_settings.proxy else unset
+    users_whitelist = (
+        f"<blue>{','.join(telegram_settings.users_whitelist)}</blue>" if telegram_settings.users_whitelist else unset
     )
-    gpt4_state = "<green>ENABLED</green>" if gpt_settings.gpt4_enabled else "<red>DISABLED</red>"
-    storage = "<red>REDIS</red>" if application_settings.redis else "<blue>LOCAL</blue>"
+    groups_whitelist = (
+        f"<blue>{telegram_settings.groups_whitelist}</blue>" if telegram_settings.groups_whitelist else unset
+    )
+    images_whitelist = (
+        f"<blue>{','.join(gpt_settings.image_generations_whitelist)}</blue>"
+        if gpt_settings.image_generations_whitelist
+        else unset
+    )
 
     messages = (
         f"Application is initialized in the {mode} mode using {storage} storage.",
+        f"Anthropic (Claude-3) client set: {is_set if bool(gpt_settings.anthropic_key) else unset }",
+        f"Mistral AI client: {is_set if bool(gpt_settings.mistralai_key) else unset }",
+        f"OpenAI client set: {is_set if bool(gpt_settings.openai_key) else unset }",
         f"Bot name is <blue>{telegram_settings.bot_name}</blue>",
         f"Initial assistant prompt: <blue>{gpt_settings.assistant_prompt}</blue>",
-        f"Access to GPT-4 models is {gpt4_state}.",
-        f"Proxy is <blue>{telegram_settings.proxy or 'UNSET'}</blue>",
+        f"Proxy is {proxy}",
         f"Messages TTL: <blue>{gpt_settings.max_conversation_age_minutes} minutes</blue>",
         f"Maximum conversation history size: <blue>{gpt_settings.max_history_tokens}</blue> tokens",
         f"Maximum answer size: <blue>{gpt_settings.max_tokens}</blue> tokens",
-        f"GPT-4 access whitelist: <blue>{gpt_settings.gpt4_whitelist or 'UNSET'}</blue>",
-        f"Users whitelist: <blue>{telegram_settings.users_whitelist or 'UNSET'}</blue>",
-        f"Groups whitelist: <blue>{telegram_settings.groups_whitelist or 'UNSET'}</blue>",
-        f"Images generation limit: <blue>{gpt_settings.image_generations_monthly_limit or 'UNSET'}</blue>",
-        f"Images limit whitelist: <blue>{gpt_settings.image_generations_whitelist or 'UNSET'}</blue>",
+        f"Users whitelist: {users_whitelist}",
+        f"Groups whitelist: {groups_whitelist}",
+        f"Images generation limit: <blue>{gpt_settings.image_generations_monthly_limit}</blue>",
+        f"Images limit whitelist: {images_whitelist}",
     )
     for message in messages:
         logger.opt(colors=True).info(message)